+++ sdk/monitor/azure-monitor-query/tests/test_metrics_client_async.py
@@ -52,3 +52,12 @@ class TestMetricsClientAsync(MetricsClientTestCase):
                 assert metric.timeseries
                 for t in metric.timeseries:
                     assert t.metadata_values is not None
+
+    @pytest.mark.asyncio
+    async def test_client_different_endpoint(self):
+        credential = self.get_credential(MetricsClient)
+        endpoint = "https://usgovvirginia.metrics.monitor.azure.us"
+        client = MetricsClient(endpoint, credential)
+
+        assert client._endpoint == endpoint
+        assert "https://metrics.monitor.azure.us/.default" in client._client._config.authentication_policy._scopes